LPS-26118 Additional Cleanup in LayoutExporter

+++ portal-impl/src/com/liferay/portal/lar/LayoutExporter.java
@@ -29,6 +29,7 @@ import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.kernel.util.StreamUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -83,6 +84,7 @@ import com.liferay.util.ContentUtil;
 import de.schlichtherle.io.FileInputStream;
 
 import java.io.File;
+import java.io.InputStream;
 
 import java.util.Date;
 import java.util.HashSet;
@@ -332,19 +334,27 @@ public class LayoutExporter {
 				String path = getLayoutSetPrototype(
 					portletDataContext, layoutSetPrototypeUuid);
 
-				File layoutSetPrototypeFile =
-					SitesUtil.exportLayoutSetPrototype(
-						layoutSetPrototype, serviceContext);
+				File layoutSetPrototypeFile = null;
+
+				InputStream inputStream = null;
 
 				try {
+					layoutSetPrototypeFile = SitesUtil.exportLayoutSetPrototype(
+						layoutSetPrototype, serviceContext);
+
+					inputStream = new FileInputStream(layoutSetPrototypeFile);
+
 					portletDataContext.addZipEntry(
-						path.concat(".lar"),
-						new FileInputStream(layoutSetPrototypeFile));
+						path.concat(".lar"), inputStream);
 					portletDataContext.addZipEntry(
 						path.concat(".xml"), layoutSetPrototype);
 				}
 				finally {
-					layoutSetPrototypeFile.delete();
+					StreamUtil.cleanUp(inputStream);
+
+					if (layoutSetPrototypeFile != null) {
+						layoutSetPrototypeFile.delete();
+					}
 				}
 			}
 		}
